Add files via upload
extract some function from train.py and put them to io.py
